Refactored Category-System Part 1
- Splitted the account_settings_provider file
- Removed duplicate Code when accessing categories
- Added CategoryMap Class
