+++ lib/constants/settings_enums.dart
@@ -1,0 +1,37 @@
+enum StandardCategoryIncome {
+  none,
+  income,
+  allowance,
+  sideJob,
+  investments,
+  childSupport,
+  interest,
+  miscellaneous,
+}
+
+enum StandardCategoryExpense {
+  none,
+  food,
+  freeTime,
+  house,
+  lifestyle,
+  car,
+  miscellaneous,
+}
+
+enum StandardAccount {
+  none,
+  debit,
+  credit,
+  cash,
+  depot,
+}
+
+enum RepeatDuration {
+  none,
+  daily,
+  weekly,
+  monthly,
+  // TODO implement custom repeat
+  // custom,
+}
